requirements.txt
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.3
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
albumentations>=1.3.1 